refactor(Icons): :recycle: update typings to use images instead of icons
