--- src/typings/icons.ts
@@ -1,3 +1,0 @@
-export interface IImage {
-  forceColor?: string;
-}
