+++ src/typings/images.ts
@@ -1,0 +1,5 @@
+import { IGenericComponent } from "./components";
+
+export interface IImage extends IGenericComponent {
+  forceColor?: string;
+}
